+++ assert_test.go
@@ -4,8 +4,11 @@ import (
 	"errors"
 	"fmt"
 	"go/types"
+	"io/ioutil"
 	"math/rand"
+	"os"
 	"reflect"
+	"regexp"
 	"testing"
 	"time"
 
@@ -1072,4 +1075,165 @@ func TestAssertNotSameElementsFails(t *testing.T) {
 			AssertNotSameElements(t, s, ss)
 		})
 	})
+}
+
+func TestAssertRegexp(t *testing.T) {
+	AssertRegexp(t, "p([a-z]+)ch", "peache")
+	rxp, err := regexp.Compile("Hello, .*")
+	AssertNoError(t, err)
+	AssertRegexp(t, rxp, "Hello, World!")
+}
+
+func TestAssertRegexp_fail(t *testing.T) {
+	AssertTestFails(t, func(t TestingPackageWithFailFunctions) {
+		AssertRegexp(t, "p([a-z]+)ch", "peahe")
+		rxp, err := regexp.Compile("Hello, .*")
+		AssertNoError(t, err)
+		AssertRegexp(t, rxp, "Hello World!")
+	})
+}
+
+func TestAssertNotRegexp(t *testing.T) {
+	AssertNotRegexp(t, "p([a-z]+)ch", "peahe")
+	rxp, err := regexp.Compile("Hello, .*")
+	AssertNoError(t, err)
+	AssertNotRegexp(t, rxp, "Hello World!")
+}
+
+func TestAssertNotRegexp_fail(t *testing.T) {
+	AssertTestFails(t, func(t TestingPackageWithFailFunctions) {
+		AssertNotRegexp(t, "p([a-z]+)ch", "peache")
+		rxp, err := regexp.Compile("Hello, .*")
+		AssertNoError(t, err)
+		AssertNotRegexp(t, rxp, "Hello, World!")
+	})
+}
+
+func TestAssertFileExists(t *testing.T) {
+	t.Run("LICENSE", func(t *testing.T) {
+		AssertFileExists(t, "LICENSE")
+	})
+
+	t.Run("README.md", func(t *testing.T) {
+		AssertFileExists(t, "README.md")
+	})
+
+	t.Run("CODE_OF_CONDUCT.md", func(t *testing.T) {
+		AssertFileExists(t, "CODE_OF_CONDUCT.md")
+	})
+
+	t.Run("CONTRIBUTING.md", func(t *testing.T) {
+		AssertFileExists(t, "CONTRIBUTING.md")
+	})
+
+	t.Run("CHANGELOG.md", func(t *testing.T) {
+		AssertFileExists(t, "CHANGELOG.md")
+	})
+}
+
+func TestAssertFileExists_fail(t *testing.T) {
+	t.Run("asdasdasdasd.md", func(t *testing.T) {
+		AssertTestFails(t, func(t TestingPackageWithFailFunctions) {
+			AssertFileExists(t, "asdasdasdasd.md")
+		})
+	})
+}
+
+func TestAssertNoFileExists(t *testing.T) {
+	t.Run("asdasdasdasd.md", func(t *testing.T) {
+		AssertNoFileExists(t, "asdasdasdasd.md")
+	})
+}
+
+func TestAssertNoFileExists_fail(t *testing.T) {
+	t.Run("LICENSE", func(t *testing.T) {
+		AssertTestFails(t, func(t TestingPackageWithFailFunctions) {
+			AssertNoFileExists(t, "LICENSE")
+		})
+	})
+
+	t.Run("README.md", func(t *testing.T) {
+		AssertTestFails(t, func(t TestingPackageWithFailFunctions) {
+			AssertNoFileExists(t, "README.md")
+		})
+	})
+
+	t.Run("CODE_OF_CONDUCT.md", func(t *testing.T) {
+		AssertTestFails(t, func(t TestingPackageWithFailFunctions) {
+			AssertNoFileExists(t, "CODE_OF_CONDUCT.md")
+		})
+	})
+
+	t.Run("CONTRIBUTING.md", func(t *testing.T) {
+		AssertTestFails(t, func(t TestingPackageWithFailFunctions) {
+			AssertNoFileExists(t, "CONTRIBUTING.md")
+		})
+	})
+
+	t.Run("CHANGELOG.md", func(t *testing.T) {
+		AssertTestFails(t, func(t TestingPackageWithFailFunctions) {
+			AssertNoFileExists(t, "CHANGELOG.md")
+		})
+	})
+}
+
+func TestAssertDirExists(t *testing.T) {
+	for _, dir := range []string{"ci", "internal", "testdata", os.TempDir()} {
+		t.Run(dir, func(t *testing.T) {
+			AssertDirExist(t, dir)
+		})
+	}
+}
+
+func TestAssertDirExists_fail(t *testing.T) {
+	for _, dir := range []string{"asdasdasdasd", "LICENSE", "CHANGELOG.md"} {
+		t.Run(dir, func(t *testing.T) {
+			AssertTestFails(t, func(t TestingPackageWithFailFunctions) {
+				AssertDirExist(t, dir)
+			})
+		})
+	}
+}
+
+func TestAssertNoDirExists(t *testing.T) {
+	for _, dir := range []string{"asdasdasdasd", "LICENSE", "CHANGELOG.md"} {
+		t.Run(dir, func(t *testing.T) {
+			AssertNoDirExists(t, dir)
+		})
+	}
+}
+
+func TestAssertNoDirExists_fail(t *testing.T) {
+	for _, dir := range []string{"ci", "internal", "testdata", os.TempDir()} {
+		t.Run(dir, func(t *testing.T) {
+			AssertTestFails(t, func(t TestingPackageWithFailFunctions) {
+				AssertNoDirExists(t, dir)
+			})
+		})
+	}
+}
+
+func TestAssertDirEmpty(t *testing.T) {
+	tempdir, _ := ioutil.TempDir("testdata", "temp")
+	AssertDirEmpty(t, tempdir)
+	defer os.RemoveAll(tempdir)
+}
+
+func TestAssertDirEmpty_fail(t *testing.T) {
+	AssertTestFails(t, func(t TestingPackageWithFailFunctions) {
+		AssertDirEmpty(t, "non-existent")
+		AssertDirEmpty(t, "internal")
+	})
+}
+
+func TestAssertDirNotEmpty(t *testing.T) {
+	AssertDirNotEmpty(t, "testdata")
+}
+
+func TestAssertDirNotEmpty_fail(t *testing.T) {
+	tempdir, _ := ioutil.TempDir("testdata", "temp")
+	AssertTestFails(t, func(t TestingPackageWithFailFunctions) {
+		AssertDirNotEmpty(t, tempdir)
+	})
+	defer os.RemoveAll(tempdir)
 }