+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 <a name="unreleased"></a>
 ## [Unreleased]
 
-### Features
-- proposal AssertSameElements
+### Bug Fixes
+- **assert:** fixed missing custom messages in some functions
 
 
 <a name="v0.2.12"></a>